+++ react/ProductSearchContextProvider.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
 
+import ProductSearchDataLayer from './components/ProductSearchDataLayer'
 import SearchQueryContainer from './components/SearchQueryContainer'
 import { searchContextPropTypes } from './constants/propTypes'
 import { SearchQueryContext } from './constants/searchContext'
@@ -11,18 +12,26 @@ class ProductSearchContextProvider extends Component {
     const props = {
       ...this.props,
       // todo: this logic should be in SearchQueryContainer
-      ...this.props.params.brand && {
+      ...(this.props.params.brand && {
         query: {
           ...this.props.query,
           map: this.props.map || 'b',
         },
-      },
+      }),
     }
 
     return (
       <SearchQueryContainer {...props}>
         <SearchQueryContext.Consumer>
-          {contextProps => React.cloneElement(this.props.children, contextProps)}
+          {contextProps => (
+            <ProductSearchDataLayer
+              searchQuery={contextProps.searchQuery}
+              loading={
+                contextProps.state.loading || contextProps.searchQuery.loading
+              }>
+              {React.cloneElement(this.props.children, contextProps)}
+            </ProductSearchDataLayer>
+          )}
         </SearchQueryContext.Consumer>
       </SearchQueryContainer>
     )